+++ android/app/src/main/res/drawable/launch_background.xml
@@ -3,7 +3,7 @@
 <layer-list xmlns:android="http://schemas.android.com/apk/res/android">
     <item android:drawable="@android:color/white" />
 
-    <!-- You can insert your own image assets here -->
+    <!-- You can insert your own image assets here --> 
     <!-- <item>
         <bitmap
             android:gravity="center"
