--- lib/cardFilter.dart
@@ -1,41 +1,0 @@
-//Smörja som förmodligen kommer tas bort
-
-
-/*import 'TodoListView.dart';
-import 'model.dart';
-
-class CardFilter {
-  List<TodoCards> _list;
-
-  CardFilter() {
-    _list = new List();
-  }
-
-  void addNewTodoCard(TodoCards todo) {
-    _list.add(todo);
-  }
-
-  void removeTodoCard(TodoCards todo) {
-    _list.remove(todo);
-  }
-
-  List<TodoCards> getAllCards() {
-    return _list;
-  }
-
-  List<TodoCards> getCheckedCards() {
-    return _list.where((card) => card.checked).toList();
-  }
-
-  List<TodoCards> getUncheckedCards() {
-    return _list.where((card) => card.checked).toList();
-  }
-
-  List<TodoCards> getAllSorted() {
-    List<TodoCards> mixed = getUncheckedCards();
-    mixed.addAll(getCheckedCards());
-    return mixed;
-  }
-}
-
-*/
